+++ Sources/CodexBarCore/WidgetSnapshot.swift
@@ -8,7 +8,6 @@ public struct WidgetSnapshot: Codable, Sendable {
         public let secondary: RateWindow?
         public let tertiary: RateWindow?
         public let creditsRemaining: Double?
-        public let codeReviewRemainingPercent: Double?
         public let tokenUsage: TokenUsageSummary?
         public let dailyUsage: [DailyUsagePoint]
 
@@ -19,7 +18,6 @@ public struct WidgetSnapshot: Codable, Sendable {
             secondary: RateWindow?,
             tertiary: RateWindow?,
             creditsRemaining: Double?,
-            codeReviewRemainingPercent: Double?,
             tokenUsage: TokenUsageSummary?,
             dailyUsage: [DailyUsagePoint])
         {
@@ -29,7 +27,6 @@ public struct WidgetSnapshot: Codable, Sendable {
             self.secondary = secondary
             self.tertiary = tertiary
             self.creditsRemaining = creditsRemaining
-            self.codeReviewRemainingPercent = codeReviewRemainingPercent
             self.tokenUsage = tokenUsage
             self.dailyUsage = dailyUsage
         }